% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{cellcount_table}
\alias{cellcount_table}
\title{Create cellcount table for calculating percentage parent}
\usage{
cellcount_table(x, percentTimer = TRUE)
}
\arguments{
\item{x}{A data frame containing the Timer Angle variable with the column name Angle.}

\item{percentTimer}{A logical value for whether Percent Timer data is produced. Default is FALSE and produces Percent Parent data.}
}
\value{
A data frame for the cell number of each Tocky Locus. NA is returned for Timer negative cells.
}
\description{
Create cellcount table for calculating percentage parent
}
\examples{
\dontrun{
cellcounttable <- cellcount_table(x, percentTimer = FALSE)
}
}
\keyword{internal}
