% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyLocus.R
\name{GetStatsTockyLocus}
\alias{GetStatsTockyLocus}
\title{Perform Statistical Tests for Tocky Locus Analysis}
\usage{
GetStatsTockyLocus(
  x,
  percentTimer = FALSE,
  p_adjust_method = "BH",
  method = "ASR",
  verbose = TRUE
)
}
\arguments{
\item{x}{A \code{TockyPrepData} object containing Tocky Locus data.}

\item{percentTimer}{Logical. If \code{TRUE}, the percentages of Timer-positive cells will be used; if \code{FALSE}, percentages of parent cells will be used.}

\item{p_adjust_method}{Character string specifying the method for p-value adjustment in multiple testing. Default is \code{'BH'} (Benjamini-Hochberg). Other methods available in \code{\link{p.adjust}}, such as \code{'holm'} or \code{'bonferroni'}, can also be used.}

\item{method}{Character string specifying the statistical test method to use. Options are:
\describe{
\item{\code{'Wilcox'}}{Mann-Whitney U test (Wilcoxon rank sum test) without data transformation.}
\item{\code{'ASR'}}{Arcsine Square Root Transformation, followed by a normality test and t-test.}
\item{\code{'Logit'}}{Logit Transformation, followed by a normality test and t-test.}
}}

\item{verbose}{Logical indicating whether to print progress messages and outputs. Default is \code{TRUE}.}
}
\value{
A \code{TockyPrepData} object containing the statistical outputs for Tocky Locus Analysis, stored in \code{x@Tocky$TockyLocusStats}.
}
\description{
This function performs statistical tests on Tocky Locus data, allowing for different methods and p-value adjustments.
}
\examples{
\dontrun{
x <- GetStatsTockyLocus(x, method = 'ASR')
}
}
