% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{locus_mean}
\alias{locus_mean}
\title{Calculate mean percentage for each locus using cellcount_table}
\usage{
locus_mean(x, sampledef = sampledef)
}
\arguments{
\item{x}{A TockyPrepData or the output of cellcount_table}

\item{sampledef}{a sample_definition data frame}
}
\value{
A summary table as a data frame.
}
\description{
Calculate mean percentage for each locus using cellcount_table
}
\examples{
\dontrun{
p <- locus_mean(x)
}
}
\keyword{internal}
