% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tocky_locus.R
\name{plot_tocky_locus}
\alias{plot_tocky_locus}
\title{Generate basic QC plots for Tocky data (Timer-Blue vs Timer-Red 2d plots)}
\usage{
plot_tocky_locus(
  x,
  file = "PlotTockyLocus",
  n = 3,
  max_cell_number = 20000,
  viridis = FALSE,
  interactive = FALSE
)
}
\arguments{
\item{x}{A TockyPrepData object produced by the function \code{prep_tocky}.}

\item{file}{The name of the output file.}

\item{n}{The number of plots per row and column in the output grid.}

\item{max_cell_number}{The maximum number of cells to be displayed per panel.}

\item{viridis}{(Optional). If TRUE, a colour-blind friendly colour set is used.}

\item{interactive}{(Optional). If TRUE, an interactive session is used to trim plot area.}
}
\value{
An unchanged TockyPrepData object, primarily for consistency in pipeline usage.
}
\description{
This function creates quick control plots for the TockyPrepData object
analyzing fluorescence changes over time in cellular activities.
}
\examples{
\dontrun{
plot_tocky_locus(data)
}
}
