% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyLocus.R
\name{plotAngleDensity}
\alias{plotAngleDensity}
\title{Plot Density of Angles by Group Using Ridge Plots}
\usage{
plotAngleDensity(x, alpha = 0.3, group_order = NULL, scale = 2, legend = FALSE)
}
\arguments{
\item{x}{A TockyPrepData object that has been processed with the \code{timer_transform} function.}

\item{alpha}{A number between 0 and 1 to be usedby ggridges.}

\item{group_order}{Optional. A character vector to define the order of group}

\item{scale}{A scaling factor to scale the height of the ridgelines. Used by ggridges.}

\item{legend}{Logical. If TRUE, legend is included.}
}
\value{
A ggplot object showing the density distribution of angles by group.
}
\description{
This function takes a TockyPrepData object, which should have been previously
processed using the \code{timer_transform} function, and creates a ridge plot showing
the density distribution of angles for each group defined in the dataset.
}
\examples{
\dontrun{
plotAngleDensity(x)
}

}
