% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyLocus.R
\name{plotTockyLocus}
\alias{plotTockyLocus}
\title{Produce scatter plots of percentages of cells in each Tocky Locus.}
\usage{
plotTockyLocus(
  x,
  percentTimer = FALSE,
  group_order = NULL,
  locus_colours = NULL,
  group_colors = NULL,
  group_by = TRUE,
  p_adjust_method = "fdr",
  ylim = NULL,
  stats = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{A TockyPrepData object}

\item{percentTimer}{A logical value for whether Percent Timer data is produced. Default is FALSE and produces Percent Parent data.}

\item{group_order}{The order of groups (optional).}

\item{locus_colours}{(optional) to choose colours for Tocky Loci.}

\item{group_colors}{(optional) to choose colours for groups.}

\item{group_by}{A logical value for whether different groups are plotted in different panels.}

\item{p_adjust_method}{A method for p-value adjustment in statistical tests.}

\item{ylim}{(Optional) the range of y values to be displayed.}

\item{stats}{A logical value for whether to produce statistical outputs. This is effective only for two-group analysis.}

\item{verbose}{Logical indicating whether to print Tocky Locus stats. Default is \code{TRUE}.}
}
\value{
A ggplot object
}
\description{
Produce scatter plots of percentages of cells in each Tocky Locus.
}
\examples{
\dontrun{
plotTockyLocus(x)
}
}
