% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tocky_locus.R
\name{plotTockyLocusLegend}
\alias{plotTockyLocusLegend}
\title{Plot Coloured Tocky Locus Legend}
\usage{
plotTockyLocusLegend(mar_par = c(4, 4, 10, 4))
}
\arguments{
\item{mar_par}{parameters for the function mar. The default is c(4, 4, 10, 4)}
}
\value{
A plot with colored rectangles and labels.
}
\description{
Plot Coloured Tocky Locus Legend
}
\examples{
\dontrun{
  plotTockyLocusLegend()
}
}
