% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyLocus.R
\name{TockyLocus}
\alias{TockyLocus}
\title{Calculate Tocky Locus using Timer Angle}
\usage{
TockyLocus(x)
}
\arguments{
\item{x}{A data.frame object or a TockyPrepData object}
}
\value{
Input data frame object including Tocky Locus data.
}
\description{
Calculate Tocky Locus using Timer Angle
}
\examples{
\dontrun{
x <- TockyLocus(x)
}

}
