% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tocky_locus.R
\name{TockyLocusLegend}
\alias{TockyLocusLegend}
\title{Generate Tocky Locus Legend For a Plot}
\usage{
TockyLocusLegend(legend = TRUE, cex = 1, viridis = FALSE)
}
\arguments{
\item{legend}{A logical arguement.}

\item{cex}{A numeric value for the text size.}

\item{viridis}{(Optional). If TRUE, a colour-blind friendly colour set is used.}
}
\description{
Generate Tocky Locus Legend For a Plot
}
\examples{
\dontrun{
TockyLocusLegend()
}
}
